AI Manager and benchmark and regresion skills updated
Split the AI Manager into a AI manager only test Synthetic test and a
combined test with Collisions and pathfinding. This way we can test
regresions in the AI Manager itself with processing and or combined
regression. May split out the Collision and Pathfinding as seperate
tests just so we can pin point specific regressions in the chain. This
will be a later update. This showcased the modlarity of the Singletons
for this use case as we can measure what we need to find true
regressions!! Also, updated claudes regression and benchmark skills to
reflect the changes.